Refactor using custom hooks
