+++ mini-redux/src/useAppState.ts
@@ -1,4 +1,4 @@
-import { createContext } from "react";
+import { createContext, useEffect, useReducer } from "react";
 import { useAppStateStore } from "./useAppStateStore";
 
 export interface AppState {
@@ -13,6 +13,14 @@ export const appStateContext = createContext<AppState | null>(null);
 
 export function useAppState() {
   const store = useAppStateStore();
+
+  const [, forceUpdate] = useReducer((v) => v + 1, Number.MIN_SAFE_INTEGER);
+  useEffect(() => {
+    const unsubscribe = store.subscribe(forceUpdate);
+
+    return unsubscribe;
+  }, [store]);
+
   const appState = store.getState();
 
   return appState;
